Add NOAHRES to restart files.
Possible inconsistency in output variables regarding the NOAH surface budget. Indeed, NOAHRES is only included in `history` when other members of the budget are included in both `history` and `restart`.

As a consequence, NOAHRES is 0 at every imulation restart. This makes it quite hard to track the surface energy balance. As it is only a supplementary 2D variable, adding it to the `restart` should only change unsignificantly in terms of storage.